agri-microservice/krishibot/requirements.txt
fastapi
uvicorn
transformers
torch
gtts
sentence-transformers
faiss-cpu
google-genai
langchain
langchain-community
langchain-huggingface
langchain-core
python-multipart
python-dotenv
pydub
librosa
tqdm
watchfiles
